move all logic to AboutActivity and add circular progress icon
