+++ app/src/main/res/drawable/ic_about_up_to_date.xml
@@ -1,0 +1,1 @@
+<!-- drawable/check.xml --><vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:width="24dp" android:viewportWidth="24" android:viewportHeight="24"><path android:fillColor="#000000" android:pathData="M21,7L9,19L3.5,13.5L4.91,12.09L9,16.17L19.59,5.59L21,7Z" /></vector>
